Proposed updates to tardis_line_identifier.py
Here I am proposing updates to tardis_line_identifier.py. The updates enable a similar bar chart output as before, with the option now to export all line transitions in the specified wavelength range to an external file. Additionally, the calculation of the line fractions for each transition is updated - before it plotted the fraction of the transitions based on the lines included in the plot. I have corrected it so that it shows the true fraction of the line transitions based off ALL line transitions in the specified wavelength range